Aula 3

+++ server/src/routes/memories.ts
@@ -3,13 +3,21 @@ import { prisma } from "../lib/prisma"
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance) {
-
-  app.get('/memories', async () => {
+app.addHook('preHandler', async (request) => {
+  await request.jwtVerify()
+})
+ 
+ 
+  app.get('/memories', async (request) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       }
-    })
+    },
+    )
 
     return memories.map((memory) => {
       return {
@@ -21,8 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
   })
 
 
-  app.get('/memories:id', async (request) => {
-    //const { id } = request.params
+  app.get('/memories:id', async (request, reply) => {
 
     const paramsSchema = z.object({
       id: z.string().uuid(),
@@ -31,10 +38,16 @@ export async function memoriesRoutes(app: FastifyInstance) {
     const { id } = paramsSchema.parse(request.params)
 
     const memory = await prisma.memory.findUniqueOrThrow({
-      where:{
+      where: {
         id,
       }
     })
+
+    if(!memory.isPublic && memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    return memory
   })
 
   app.post('/memories', async (request) => {
@@ -44,60 +57,80 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic: z.coerce.boolean().default(false)
     })
 
-    const {content, coverUrl,isPublic} = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
     const memory = await prisma.memory.create({
-      data:{
+      data: {
         content,
         coverUrl,
         isPublic,
-        userId: '6f632fe0-025b-409c-89df-f7dc16c3084b'
+        userId: request.user.sub,
       }
     })
 
     return memory
   })
 
-  app.put('/memories:id', async (request) => {
+  app.put('/memories:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
-    
+
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
       isPublic: z.coerce.boolean().default(false)
     })
 
-    const {content, coverUrl,isPublic} = bodySchema.parse(request.body)
+    const { content, coverUrl, isPublic } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
-      where:{
+    let memory = await prisma.memory.findUnique({
+      where: {
         id,
       },
-      data:{
+    })
+
+    if(memory?.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
+      where: {
+        id,
+      },
+      data: {
         content,
         coverUrl,
-        isPublic
-      }
+        isPublic,
+      },
     })
-    
+
     return memory
   })
 
-  app.delete('/memories:id', async (request) => {
+  app.delete('/memories:id', async (request, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    await prisma.memory.delete({
-      where:{
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
         id,
       }
+    })
+
+    if(memory?.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    await prisma.memory.delete({
+      where: {
+        id,
+      },
     })
   })
 }
